requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
fasttext==0.9.2
gunicorn==20.0.4
importlib-metadata==1.6.1
isodate==0.6.0
Markdown==3.2.2
numpy==1.19.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pybind11==2.5.0
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
rdflib==5.0.0
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
zipp==3.1.0
